devuelve un parking y su listado de plazas

+++ src/main/java/com/quicklypark/back/acceso/repository/ParkingRepository.java
@@ -3,6 +3,6 @@ package com.quicklypark.back.acceso.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.quicklypark.back.acceso.entity.ParkingEntity;
 
-public interface ParkingRepository extends CrudRepository<ParkingEntity, Integer> {
+public interface ParkingRepository extends CrudRepository<ParkingEntity, Long> {
 
 }
